feat : chatting Server 연결
