+++ src/components/ChatBox.tsx
@@ -1,5 +1,9 @@
-import { useState } from 'react';
+import { useState, useRef, useEffect} from 'react';
+import SockJS from 'sockjs-client';
+import { CompatClient, Stomp } from '@stomp/stompjs';
 import useUserStore from '../store/userStore';
+
+import { CHAT_URL, getAccessToken } from '../constants/api';
 
 import styled from 'styled-components';
 import { FlexLayout } from '../styles/layout';
@@ -38,14 +42,60 @@ export default function ChatBox({ isShort, text }: ChatBoxProps) {
   const [userChatting, setUserChatting] = useState<string[][]>([]);
   const [chatting, setChatting] = useState('');
   const { userData } = useUserStore();
+  const client = useRef<CompatClient | null>(null);
+  const ACCESS_TOKEN = getAccessToken();
 
   const handleChange = (e: { target: { value: string } }) => {
     setChatting(e.target.value);
   };
 
   const handleClick = () => {
-    setUserChatting([...userChatting, [userData.nickname, chatting]]);
+    sendChatMessage(chatting);
   };
+
+  const connectChatHandler = () => {
+    const socket = new SockJS(`${CHAT_URL}/socket`);
+    const header = {
+      Authorization: `Bearer ${ACCESS_TOKEN}`,
+      'Content-Type': 'application/json',
+    };
+    client.current = Stomp.over(socket);
+    client.current.connect(header, () => {
+      client.current?.subscribe(
+        `/topic/lobby-chat`,
+        (message) => {
+          const fetchedData = JSON.parse(message.body);
+          setUserChatting(prevChatting => {
+            const newChatting = [...prevChatting, [fetchedData.nickname, fetchedData.text]];
+            return newChatting;
+          });
+        },
+        header
+      );
+    });
+  };
+
+  const sendChatMessage = (chatting: string) => {
+    const socket = new SockJS(`${CHAT_URL}/socket`);
+    const header = {
+      Authorization: `Bearer ${ACCESS_TOKEN}`,
+      'Content-Type': 'application/json',
+    };
+    const message = JSON.stringify({ text: chatting, nickname: userData.nickname, color: userData.color });
+    client.current = Stomp.over(socket);
+    client.current.connect(header, () => {
+      client.current?.send(
+        `/topic/lobby-chat`,
+        header,
+        message
+      );
+    });
+  };
+
+  useEffect(() => {
+    connectChatHandler();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [ACCESS_TOKEN]);
 
   return (
     <WhiteBox mode={isShort ? 'MEDIUM' : 'TALL'} width='32rem'>
