+++ src/pages/Lobby/Lobby.tsx
@@ -19,7 +19,6 @@ export default function LobbyPage() {
   const navigate = useNavigate();
   const ACCESS_TOKEN = getAccessToken();
   const client = useRef<CompatClient | null>(null);
-  const client2 = useRef<CompatClient | null>(null);
 
   const [isOpen, setIsOpen] = useState({
     newRoomDialog: false,
@@ -51,25 +50,6 @@ export default function LobbyPage() {
     });
   };
 
-  const connectChatHandler = () => {
-    console.log("CHAT URL" + CHAT_URL)
-    const socket = new SockJS(`${CHAT_URL}/socket`);
-    const header = {
-      Authorization: `Bearer ${ACCESS_TOKEN}`,
-      'Content-Type': 'application/json',
-    };
-    client2.current = Stomp.over(socket);
-    client2.current.connect(header, () => {
-      client2.current?.subscribe(
-        `/topic/lobby-chat`,
-        (message) => {
-          console.log("Received message:", message.body);
-        },
-        header
-      );
-    });
-  };
-
   const showRoomList = async () => {
     try {
       await getRoomlist();
@@ -79,11 +59,6 @@ export default function LobbyPage() {
     }
   };
 
-  useEffect(() => {
-    connectHandler();
-    connectChatHandler();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [ACCESS_TOKEN]);
 
   const handleCheckDialog = (id: number, isPublic: boolean) => {
     setClickedRoom({ roomId: id, isPublic: isPublic });
